Split v2.rest_framework package tests.
